Draw link follow hint outside of the link rect
Place it to the right and above for increased visibility.

Draw a background for the link hints.

Do not use highlight-transparency for the foreground transparency, but
rather the transparency of highlight-fg. (Does this make the
transparency option obsolete?)

TODO: The rectangle around the link is a bit weird. I want the same
rectangle size for links and search results as the mouse selection gets.