Do not abort failed search w/ incremental-search
Zathura displays the "Pattern not found" message when the search fails.
This aborts the search.

When incremental-search is turned on, this happens directly after
pressing a key, not just after confirming the search. This in turn leads
to all following keys being interpreted as commands/mappings to zathura.
When quickly typing a search term and mistyping it, a bunch of commands
are executed by mistake.

For now, just turn off the error message if incremental-search is turned
on.

TODO for later would be to still display the message when the search was
accepted.